5cef001d85ad066031fb4febbbca666f9eadc533 Fix: Newly added rows did not have `dtr-hidden` class added to their cells if they were immediately hidden
Fix: Responsive needs to call `-api columns.adjust()` if it has changed the column visibility. This is parimarily for scrolling so that the cloned header / footer elements can be updated.

There is a warning to this action, since it would trigger Responsive's
own resizing action on the `column-sizing` event listener. But there
will be no change the second time around. So it isn't optimal, but
unless we expose some kind of scroll draw option in DataTables, this is
as close as we can get.

Sync to source repo @5cef001d85ad066031fb4febbbca666f9eadc533